refactor: Create an AppLayout component to increase lisibility
