+++ src/components/AppRouter.jsx
@@ -6,6 +6,7 @@ import ContactImportationModal from 'components/ContactImportationModal'
 import ContactFormModal from 'components/Modals/ContactFormModal'
 import ContactInfoModal from 'components/Modals/ContactInfoModal'
 import ConfirmDeleteModal from 'components/Modals/ConfirmDeleteModal'
+import AppLayout from './AppLayout'
 
 const OutletWrapper = ({ Component }) => (
   <>
@@ -18,12 +19,14 @@ const AppRouter = () => {
   return (
     <HashRouter>
       <Routes>
-        <Route path="/" element={<OutletWrapper Component={App} />}>
-          <Route path="import" element={<ContactImportationModal />} />
-          <Route path="new" element={<ContactFormModal />} />
-          <Route path=":contactId" element={<ContactInfoModal />} />
-          <Route path=":contactId/edit" element={<ContactFormModal />} />
-          <Route path=":contactId/delete" element={<ConfirmDeleteModal />} />
+        <Route element={<AppLayout />}>
+          <Route path="/" element={<OutletWrapper Component={App} />}>
+            <Route path="import" element={<ContactImportationModal />} />
+            <Route path="new" element={<ContactFormModal />} />
+            <Route path=":contactId" element={<ContactInfoModal />} />
+            <Route path=":contactId/edit" element={<ContactFormModal />} />
+            <Route path=":contactId/delete" element={<ConfirmDeleteModal />} />
+          </Route>
         </Route>
       </Routes>
     </HashRouter>
